update formDialog and update onCloseDialog function in newDelivery context reset states
